% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeMLE.R
\name{as.data.frame.fittedRemMLE}
\alias{as.data.frame.fittedRemMLE}
\title{Convert objects of class fittedRemMLE to a data.frame}
\usage{
\method{as.data.frame}{fittedRemMLE}(x, row.names, optional, ...)
}
\arguments{
\item{x}{an object of class \code{fittedRemMLE}}

\item{row.names}{`NULL` or character vector containing row names for the data frame.}

\item{optional}{logical. If `TRUE`, setting row names and converrting column names is optional.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
a data frame containing one row with the following columns:
\itemize{
\item estimate: estimated number of individuals present in the population at the begining of the removal experiment
\item se: estandard error of the estimate
\item lcl: 95\% lower confidence interval
\item ucl: 95\% upper confidence interval
}
}
\description{
This method extends the generic `as.data.frame` function for
objects of class \code{fittedRemMLE}. It extracts the most
relevant component of the object, which is a data frame containing
the result of the model estimation of Population size.
}
\examples{
myEstimate <- as.data.frame(removeMLE(genHistRMark(c(100,80,60),3)))
}
\author{
Fer Arce
}
