% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recRemMLE.R
\name{as.data.frame.fittedRemMLERec}
\alias{as.data.frame.fittedRemMLERec}
\title{Convert objects of class fittedRemMLERec to a data.frame}
\usage{
\method{as.data.frame}{fittedRemMLERec}(x, row.names, optional, ...)
}
\arguments{
\item{x}{an object of class \code{fittedRemMLE}}

\item{row.names}{`NULL` or a character vector giving the row names for the data frame. Missing values are not allowed}

\item{optional}{logicalf `TRUE`, setting row names and converting column namesis optional.}

\item{...}{further arguments to be passeds to the generic method}
}
\value{
a data frame containing one row per recursive estimate  with the following columns:
\itemize{
\item estimate: estimated number of individuals present in the population at the begining of the removal experiment
\item se: estandard error of the estimate
\item lcl: 95\% lower confidence interval
\item ucl: 95\% upper confidence interval
}
}
\description{
This method extends the generic `as.data.frame` function for
objects of class \code{fittedRemMLERec}. It extracts the most
relevant component of the object, which is a data frame containing
the result of the recursively estimation of Population size.
}
\examples{
recEst <- as.data.frame(recRemMLE(genHistRMark(c(200, 150, 125), 3),3))
}
\author{
Fer Arce
}
