% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHistRMark.R
\name{as.data.frame.HistRMarkLong}
\alias{as.data.frame.HistRMarkLong}
\title{Convert objects of class HistRMarkLong to a data.frame}
\usage{
\method{as.data.frame}{HistRMarkLong}(x, row.names, optional, ...)
}
\arguments{
\item{x}{an object of class \code{HistRMarkLong}}

\item{row.names}{`NULL`.or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{optional}{logical. if `TRUE`, setting row names and converting column names (to syntactic names: see `make.names`) is optional}

\item{...}{further arguments to be passed to the call}
}
\value{
a data frame containig two columns. The first, named ch
    contains the removal histories, and the second, named ind,
    contain the number of individual asociated to each
    history. This naming approach matched the naming used in
    package \code{RMark}
}
\description{
This method extends the generic `as.data.frame` function for
objects of class \code{HistRMarkLong}. It extract the most
relevant component of the object, which is a data frame containing
the individual removal histories.
}
\examples{
myCaptHist <- as.data.frame(genHistRMark(c(200, 150, 125), 3))
}
\author{
Fer Arce
}
