% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCatches.R
\name{fitCatches}
\alias{fitCatches}
\title{Fitting a gam model to cumulative numbers of removed
    individuals.}
\usage{
fitCatches(Data, k = 5, sub = NULL)
}
\arguments{
\item{Data}{an object of class fittedRemMLERec}

\item{k}{smoothing parameterfot the gam function - number of knots
(It cannot be larger than n-1 occasions)}

\item{sub}{if prresent, subset of data to be be used for fitting
the gam model}
}
\value{
the object of a fitted gam model.
}
\description{
Function that defines the gam structured used.
}
\details{
This function fits a gam to a processed object. The object of this
function is to define the gam structure (mgcv model) over the
cumulative number of animals removed, aimed to smooth a curve with
will be projected in order to evaluate current and future
performance of the removal experiment in regards the aims of the
project. This function is not expected to be called by the end
user.
}
\author{
Fer Arce
}
