% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHistRMark.R
\name{genHistRMark}
\alias{genHistRMark}
\title{Fotrmatting removal data for RMark}
\usage{
genHistRMark(pop, events = NULL, msg = TRUE)
}
\arguments{
\item{pop}{Numerical vector containing a sequence of numbers representiong the total number of animals extracted per night}

\item{events}{Numeric. Number of removal events considering for futher analysis (equal or smaller than the number of trapping events)}

\item{msg}{\code{logical}. IF TRUE a short message containing the total number of individuals removed for the number of events considered is printed in the console.}
}
\value{
an object of class \code{HistRMarkLong} consisting on a list of the following objects: \cr
\itemize{
\item formatted: a data frame containing two columns
\itemize{
\item ch: individual removal histories
\item ind: number of individuals associated to each removal history (default to 1)
}
\item captures: number of removed individuals per removal event considered for formatting
}
}
\description{
This function generate individual trapping histories suitable to be analised by RMark (and therefore MARK)
}
\details{
This function takes the results of a removal experiment in its simples expression (pooled total of individuals per nigth or removal event/pass) and format it to be used by \code{RMark} package. RMark (and therefore MARK) accepts a few different types of data formats. I have chosen the long format because it will be consistent with the rest of modelling approaches offered in the package
}
\examples{
myHist <- genHistRMark(c(200, 150, 125), 3)
}
\author{
Fer Arce
}
