% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mayflies.R
\docType{data}
\name{mayflies}
\alias{mayflies}
\title{mayflies dataset example}
\format{
A numeric vector of length 5
}
\source{
Otis, David L, Kenneth P. Burnham, Gary C. White, and
    David R. Anderson. “Statistical Inference from Capture Data on
    Closed Animal Populations.” Wildlife Monographs 62 (1978):
    3–135.
}
\usage{
mayflies
}
\description{
A vector of succesive removals events of mayflies
}
\examples{
data(mayflies)
}
\keyword{datasets}
