% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MZest.R
\name{MZest}
\alias{MZest}
\title{Moran-Zippin estimator for two consecutive removal events}
\usage{
MZest(n1, n2)
}
\arguments{
\item{n1}{Number of individuals removed in the first event}

\item{n2}{Number of individuals removed in the second event}
}
\value{
a data frame containing one row with the following values:
\itemize{
\item Estimate: estimated population size
\item sd: standard error of the Estimate
\item lcl: lower value of the 95% confidence interval
\item ucl: upper value of the 95% confidence interval
}
}
\description{
This function generates an estimate of population size after two
consecutive removal events using a Moran-Zippin estimator
}
\details{
Function to estimate the size of a population subject of two
consecutive removal occasions. This function is provided for
reference, as it relies in strong asumptions like equal
catchability of individuals and representatitiviness of the
sanmple suize. Its use is strongly discouraged, and removal
experiments should always consist of at least 3 removal
events. The use of function \code{removeMLE} with at least three
sampling events should be the prefered approach as it adds more
flexibility.

A major problem of the Moran-Zippin estimator is that it requires
that the number of individuals removed in the second event must be
smaller than the number removed in the first event. Otherwise it
will return a negative population size estimate and unsensical
confidence interval values if n2 > n1 or Inf if n1 = n2.
}
\examples{
simpEst <- MZest(84,61)
}
\references{
Moran, P. A. P. (1951). A mathematical theory of
    animal trapping. Biometrica, 38, 307-311.

Zippin, C. (1956). An evaluation of the removal method of
estimating animal populations. Biometrics, 12, 163-189.

Zippin, C. (1958). The removal method of population
estimation. Journal of Wildlife Management, 22, 82-90.
}
\author{
Fer Arce
}
