% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeMLE.R
\name{print.fittedRemMLE}
\alias{print.fittedRemMLE}
\title{Printing method objects of class fittedRemMLE to the console}
\usage{
\method{print}{fittedRemMLE}(x, ...)
}
\arguments{
\item{x}{an object of class \code{fittedRemMLE}}

\item{...}{further generic arguments to be pased to the generic method}
}
\value{
prints a message in the console with the main results of the model:
\itemize{
\item estimate: estimated number of individuals present in the population at the begining of the removal experiment
\item se: estandard error of the estimate
\item lcl: 95\% lower confidence interval
\item ucl: 95\% upper confidence interval
}
}
\description{
This method extend the generic `print` function for objects of
class \code{fittedRemMLE}. It prints to the console the Estimated
population size of the population being monitored.
}
\examples{
print(removeMLE(genHistRMark(c(100,80,60),3)))
}
\author{
Fer Arce
}
