% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recRemMLE.R
\name{print.fittedRemMLERec}
\alias{print.fittedRemMLERec}
\title{Printing method objects of class fittedRemMLERec to the console}
\usage{
\method{print}{fittedRemMLERec}(x, ...)
}
\arguments{
\item{x}{an object of class \code{fittedRemMLERec}}

\item{...}{further arguments to be passed to the class}
}
\value{
prints a message in the console with the main results of the model:
\itemize{
\item estimate: estimated number of individuals present in the population at the begining of the removal experiment
\item se: estandard error of the estimate
\item lcl: 95\% lower confidence interval
\item ucl: 95\% upper confidence interval
}
}
\description{
This method extend the generic `print` function for objects of
class \code{fittedRemMLERec}. It prints to the console the Estimated
population size of the population being monitored.
}
\examples{
print(recRemMLE(genHistRMark(c(200, 150, 125), 3),3))
}
\author{
Fer Arce
}
