% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genHistRMark.R
\name{print.HistRMarkLong}
\alias{print.HistRMarkLong}
\title{Print objects of class HistRMarkLong to the console}
\usage{
\method{print}{HistRMarkLong}(x, ...)
}
\arguments{
\item{x}{and object of class \code{HistRMarkLong}}

\item{...}{further arguments to be passed to the call}
}
\value{
a message printed in the console
}
\description{
This method extends the generic `print` function for objects of
class \code{HistRMarkLong}. It print a summary of information into
the R console, preventing console overflow with long but informative
information.
}
\examples{
print(genHistRMark(c(200, 150, 125), 3))
}
\author{
Fer Arce
}
