% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projCatches.R
\name{projCatches}
\alias{projCatches}
\title{capture projection}
\usage{
projCatches(Data, proj = 1, sub = NULL, ...)
}
\arguments{
\item{Data}{an object of class 'fittedRemMLERec'.}

\item{proj}{numbers of future days to be projected.}

\item{sub}{optionalfurther passes to fitCatches and inside this function}

\item{...}{further arguments to be imposed to fitCatches (like
smoothing parameter or subset od data to fit the gam)}
}
\value{
predicted values
}
\description{
Function to fit and project a gam curve to the cumulative number of individuals removed.
}
\details{
This function fits a gam model to the provided number of
cumulative captures, and then project the model fit to near future
trapping events
}
\author{
Fer Arce
}
