% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recRemMLE.R
\name{recRemMLE}
\alias{recRemMLE}
\title{estimation of population size of removal experiments using MLE over an extreme Mtbh model.}
\usage{
recRemMLE(Data, events = NULL, recursive = TRUE)
}
\arguments{
\item{Data}{an object of class \code{HistRMarkLong}}

\item{events}{maximum number of events to be considered}

\item{recursive}{Logical. If TRUE, it will make the fuill recursive from, to.}
}
\value{
an object of class \code{fittedRemMLERec} containing a list of two elements: \cr
\itemize{
\item Data: the object of class `HistRMarkLong` used as imput to the function
\item result: The actual results of recursively fitting the removal model to the supplied Data. It is a data frame containing a row per recursive fit with the following information:
\itemize{
\item estimate: the estimated population size
\item se: standard error of the estimate
\item lcl: lower 95% confidence interval value
\item ucl: upper 95% confidence interval value
\item n.occ: number of removing events
\item NCatch: Cumulative number of individuals removed
\item propCatch: proportion of the population already removed (againts every iteration of the Estimated N  hat)
}
}
}
\description{
Function to fit recursively a closed population CMR model
}
\details{
This function fit closed capture-mark-recapture models with extreme behavioural response (complete trap-shyness)as detailed in Otis et al. (1978). This function calls function \code{removeMLE} recursively an generate a sequence of estimates to see the evolution and estabilization of the Population size estimations and the relative shrinking of the confidence interval.
}
\examples{
exPop <- recRemMLE(genHistRMark(c(200, 150, 125), 3),3)
}
\references{
Otis, D. L., Burnham, K. P., White, G. C., and Anderson, D. R. (1978). Statistical inference from capture data on closed animal populations. *Wildlife monographs*, 62, 1-135.
}
\author{
Fer Arce
}
