% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeMLE.R
\name{removeMLE}
\alias{removeMLE}
\title{estimation of population size of removal experiments using MLE over an extreme Mtbh model.}
\usage{
removeMLE(Data)
}
\arguments{
\item{Data}{an object of class \code{HistRMarkLong}}
}
\value{
an object of class \code{fittedRemMLE} containing a list of two elements: \cr
\itemize{
\item Data: the object of class `HistRMarkLong` used as imput to the function
\item result: The actual results of fitting the removal model to the supplied Data. It is a single row data frame containing the following information:
\itemize{
\item estimate: the estimated population size
\item se: standard error of the estimate
\item lcl: lower 95% confidence interval value
\item ucl: upper 95% confidence interval value
}
}
}
\description{
Function to fit a closed population CMR model
}
\details{
This function fit a closed capture-mark-recapture model with extreme behavioural response (complete trap-shyness). The rationale of this approach is that an extreme trap-shyness behavior will result in capture histories consisting on only one capture per individual and thus becoming a removal experiment (either released or removed, animals won't occupy the traps again -at least during the survery term). This approach is suggested in Otis et al. (1978) and has been extensively tested via simulations.

This method fit a multinomial Mtbh model implenmented in MARK and called triough RMak package. Instructions on how to install RMark can be found in the PhiDot website http://www.phidot.org
}
\examples{
exPop <- removeMLE(genHistRMark(c(100,80,60),3))
}
\references{
Otis, D. L., Burnham, K. P., White, G. C., and Anderson, D. R. (1978). Statistical inference from capture data on closed animal populations. *Wildlife monographs*, 62, 1-135.
}
\author{
Fer Arce
}
