% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remRats.R
\name{remRats}
\alias{remRats}
\title{remRats}
\usage{
remRats(pob1)
}
\arguments{
\item{pob1}{a numeric vector contianing the number of individuals removed during a remioval experiment at each event.}
}
\value{
a list from recRemMLE
}
\description{
Wrapper function to make a full process of removal data using a single function call
}
\details{
This is a wrapping function to facilitate the visualization of the analytical procedures implemented in remRats
}
\author{
Fer Arce
}
