% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remSim.R
\name{remSim}
\alias{remSim}
\title{simple removal simulation}
\usage{
remSim(N, p, j, max.capt)
}
\arguments{
\item{N}{Initial population size}

\item{p}{capturability (probabilidad of copture, equal for each individual, in the interval (0,1))}

\item{j}{number of removal events}

\item{max.capt}{maximum number of individuals potentially removed per event (i.e number of single catch traps)}
}
\value{
a vector of total number of captures per removal event
}
\description{
Function to simulate removal datasets
}
\details{
Simple simulation of removal experiments under basic constraints, following a binomial secuence of removals, with the posibility of adding a cap up to simulate experiments where trap saturation occur. To prevent this behavior to occur, max.capt should be fixed to a ridiculously unrealistic large value. This function does not allow to simulate populations  with time or individual-level varying parameters. It is defined and mantained for the purpose of simple checking and large simulations.
}
\examples{
simEx <- remSim(100, 0.5, 4, 50)
}
\author{
Fer Arce
}
