% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyView.R
\name{shinyView}
\alias{shinyView}
\title{shinyView}
\usage{
shinyView(popEst)
}
\arguments{
\item{popEst}{an object of class 'fittedRemMLERec'.}
}
\value{
open a shiny app in a web browser to visualize the data analysis and projections
}
\description{
Function to visualice the outcome of a remRats analysis
}
\details{
This function open a shiny app where the different output tables and time series plots generated are shown. This Shiny app is interactive so the end user can decider project or no the model into nearby future to avaluate the expectations of future field effort. This function shouldf be called from inside the call of remRats.
}
\author{
Fer Arce
}
