% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitefish.R
\docType{data}
\name{whitefish}
\alias{whitefish}
\title{whitefish dataset example}
\format{
A numeric vector of length 7
}
\source{
Otis, David L, Kenneth P. Burnham, Gary C. White, and David
    R. Anderson. “Statistical Inference from Capture Data on Closed
    Animal Populations.” Wildlife Monographs 62 (1978): 3–135.
}
\usage{
whitefish
}
\description{
A vector of succesive removals events of whitefish
}
\examples{
data(whitefish)
}
\keyword{datasets}
